+++ all-is-cubes/src/space/space_txn.rs
@@ -791,8 +791,8 @@ mod tests {
     use crate::behavior::NoopBehavior;
     use crate::block::{self, AIR};
     use crate::content::make_some_blocks;
-    use crate::inv::EphemeralOpaque;
     use crate::transaction::TransactionTester;
+    use crate::universe::EphemeralOpaque;
     use alloc::vec;
     use core::sync::atomic::{AtomicU32, Ordering};
     use pretty_assertions::assert_eq;